requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.1.0+cpu
torchvision==0.16.0+cpu
opencv-python-headless==4.8.1.78
pillow==10.2.0
numpy==1.24.4